### 2024-01-06 #155 @rcaneill
xnemogcm
as discussed in #154 I make a full submission

Submitting Author: Romain Caneill (@rcaneill)
Package Name: xnemogcm
One-Line Description of Package: Interface to open NEMO global circulation model output dataset with xarray and create a xgcm grid.
Repository Link: https://github.com/rcaneill/xnemogcm/
Version submitted:  0.4.2
Version accepted: TBD

---

xnemogcm is an interface to open NEMO ocean global circulation model output dataset and create a xgcm grid. NEMO 3.6, 4.0, and 4.2.0 are tested and supported. It can handle large simulations, is aware of meshgrid files, and makes it easy to handle the  netCDF outputs od NEMO.

xnemogcm extracts the NEMO output and adds metadata (+ do some other things as change coordinate names, etc), and produces new xarray datasets. It is thus data processing.

The target audience is anyone working with NEMO outputs and python. The scientific applications are any analyse that one want to do with NEMO outputs.

Yes, xorca. My package differs as 1) it is more recent, updated to work with newer versions of NEMO, and 2) it uses a very different method to sort variables on the proper grid point (center, face, etc) (cf https://xgcm.readthedocs.io/en/latest/grids.html). xorca uses hardcoded variables, while xnemogcm uses attributes (either output directly by NEMO, or they can also be given while calling the processing functions).

I made a pre-submission enquiry, issue #154
