### 2024-01-11 #155 @rcaneill
Hi @ocefpaf,
thanks for taking care of this review :)

I can directly mention that some requirements are not fulfilled:
* API documentation: not done
* CONTRIBUTING.md: not done
* CODE_OF_CONDUCT.md: not done
* Issue Submission Documentation: not done

As far as I can say, the other requirements should be fulfilled.
