### 2024-01-11 29a3d69
fix nbviewer links
