### 2024-01-11 #155 @ocefpaf
That's awesome! Here are a few minor comments/questions I have after the first pass:

- You mentioned that the GH repo is a mirror of a Gitlab one. IMO this dual repo usually leads to abandoned PRs in one place and/or stale code somewhere. Let alone the confusion to those finding the Software via search engines. I'm not against this setup and sometimes there are strong reasons to do so. However, if that can be avoided, it usually provides a smoother experience to all.
- Are you publishing the docs somewhere? I could not find them. I understand that the API docs are not ready but you do have some examples that can be the first form of user facing docs.
- I started checking the boxes above. Please let me know if I missed one that is already done.
