### 2024-01-12 rcaneill/xnemogcm#72 @rcaneill
Change default branch from master to main
https://www.theserverside.com/feature/Why-GitHub-renamed-its-master-branch-to-main

It is now a new habit within the git community to avoid the term "master", and prefer "main".
