### 2024-01-13 #155 @rcaneill
> @rcaneill we usually only start the review after all those points are checked. The reason is b/c the reviewers will read those docs and give you some feedback. Let me know if you need help setting docs, examples, etc in your repo. Once we are done I'll find two reviewers for your project and start the review process.

Ok I understand, I will try to check all boxes in the following weeks!
