+++ docs/index.rst
@@ -1,20 +1,0 @@
-.. xnemogcm documentation master file, created by
-   sphinx-quickstart on Fri Feb  7 09:48:55 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to xnemogcm's documentation!
-====================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
