+++ .readthedocs.yaml
@@ -16,16 +16,10 @@ build:
       - curl -sSL https://install.python-poetry.org | python3 -
       # Tell poetry to not use a virtual environment
       - poetry config virtualenvs.create false
-      - export PATH=$PATH:$HOME/.local/bin
     post_install:
       # Install dependencies with 'docs' dependency group
       # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
-      - poetry install --with docs
+      - $HOME/.local/bin/poetry install --with docs
 
 mkdocs:
   configuration: mkdocs.yml
-
-# Optionally declare the Python requirements required to build your docs
-python:
-  install:
-  - requirements: docs/requirements.txt
