+++ .readthedocs.yaml
@@ -15,7 +15,7 @@ build:
       # https://python-poetry.org/docs/#installing-with-the-official-installer
       - curl -sSL https://install.python-poetry.org | python3 -
       # Tell poetry to not use a virtual environment
-      - poetry config virtualenvs.create false
+      - $HOME/.local/bin/poetry config virtualenvs.create false
     post_install:
       # Install dependencies with 'docs' dependency group
       # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
