+++ .readthedocs.yaml
@@ -19,6 +19,8 @@ build:
     post_install:
       # Install dependencies with 'docs' dependency group
       # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
+      - which python
+      - which python3
       - $HOME/.local/bin/poetry install --with docs
 
 mkdocs:
