+++ .readthedocs.yaml
@@ -19,9 +19,8 @@ build:
     post_install:
       # Install dependencies with 'docs' dependency group
       # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
-      - python --version
-      - python3 --version
-      - $HOME/.local/bin/poetry install --with docs
+      - python -m pip install --requirement <($HOME/.local/bin/poetry export --with docs --format requirements.txt)
+      - python -m pip install --no-deps .
 
 mkdocs:
   configuration: mkdocs.yml
