+++ docs/new.md
@@ -1,5 +1,9 @@
 # What's new
 
+### (unreleased)
+* Rework doc: set up readthedocs
+* Drop python 3.8
+
 ### v0.4.2 (2023-08-07)
 * Allow additional dimension names occurring when variables on inner grid are diagnosed, e.g. `x_grid_U_inner` or `x_grid_U`.
 * Add coordinates into the DataArrays
