### 2024-02-05 8a860ac
Create CODE_OF_CONDUCT.md
