### 2024-02-05 11a7c56
add code of conduct in doc
