### 2024-02-08 #155 @rcaneill
> @rcaneill there seems to be a documentation build problem in page https://xnemogcm.readthedocs.io/en/latest/examples/recombing_mesh_mask_domain_cfg/

thanks for catching this, it is now corrected.

> Also, some of your syntax there is not Windows friendly and users that copy-n-paste may find themselves with errors. Those are not blockers for us to continue but I wanted to give that feedback here before I forget.

Are you speaking about the `ls` commands in the notebooks? If yes, I can change and use python instead.
