### 2024-02-08 #155 @ocefpaf
> Are you speaking about the `ls` commands in the notebooks? If yes, I can change and use python instead.

That and I believe some of the paths may not be parsed correctly. I don't have a Windows machine to test in `Path` is smart enough to convert `/` to `\` for us.
