### 2024-02-08 #155 @rcaneill
pathlib.Path is handling the windows conversion: https://stackoverflow.com/questions/2953834/how-should-i-write-a-windows-path-in-a-python-string-literal/68234511#68234511

For the `ls`, I changed to `os.listdir`
