### 2024-02-13 #155 @ocefpaf
:wave: Hi [Callum Rollo](https://github.com/callumrollo) and
[Paige Martin](https://github.com/paigem)!
Thank you for volunteering to review for pyOpenSci!

Please get in touch with any questions or concerns! Your review is due: March 5th, 2024 (3 weeks).
Please let us know if you need more time.

---

Reviewers: @callumrollo and @paigem
Due date: 2024-02-29

PS: You can comment, open issue, and/or PR directly in the [xnemogcm](https://github.com/rcaneill/xnemogcm/) for this review. Just make sure to cross-reference them here.
