### 2024-02-14 #155 @callumrollo
Hej Romain, I'm starting my review today. The package looks really good!

I see you have an Issue in to rename master >> main. Perhaps now is a good time to do that before I queue up a couple of small PRs https://github.com/rcaneill/xnemogcm/issues/72
