### 2024-02-14 9ca4801
refactor: remove unused imports and variables
