+++ .pre-commit-config.yaml
@@ -1,0 +1,35 @@
+exclude: .github
+
+repos:
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v4.5.0
+  hooks:
+    - id: trailing-whitespace
+    - id: check-ast
+    - id: debug-statements
+    - id: end-of-file-fixer
+    - id: check-docstring-first
+    - id: check-added-large-files
+
+- repo: https://github.com/psf/black
+  rev: 24.1.1
+  hooks:
+  - id: black
+    language_version: python3
+
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.2.0
+  hooks:
+    - id: ruff
+
+- repo: https://github.com/codespell-project/codespell
+  rev: v2.2.6
+  hooks:
+    - id: codespell
+      exclude: poetry.lock
+
+- repo: https://github.com/tox-dev/pyproject-fmt
+  rev: "1.7.0"
+  hooks:
+    - id: pyproject-fmt
+      additional_dependencies: ["tox>=4.9"]
