### 2024-02-14 0925306
Add installation instruction to README
