### 2024-02-14 #155 @rcaneill
Thanks @callumrollo for all the PR and improvements.

@ocefpaf am I supposed to check / merge the PR on the fly, or wait for the end of the review?
