### 2024-02-14 #155 @ocefpaf
> @ocefpaf am I supposed to check / merge the PR on the fly, or wait for the end of the review?

Up to you. Sometimes the second reviewer can add upon the first one and waiting makes sense. Sometimes merging/solving everything now makes it easier b/c it reduces the context for the second review.
