+++ .pre-commit-config.yaml
@@ -16,6 +16,7 @@ repos:
   hooks:
   - id: black
     language_version: python3
+    exclude: xnemogcm/__init__.py
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
   rev: v0.2.0
@@ -26,7 +27,7 @@ repos:
   rev: v2.2.6
   hooks:
     - id: codespell
-      exclude: poetry.lock
+      exclude: poetry.lock|namelist_ref
 
 - repo: https://github.com/tox-dev/pyproject-fmt
   rev: "1.7.0"
