+++ .pre-commit-config.yaml
@@ -16,12 +16,12 @@ repos:
   hooks:
   - id: black
     language_version: python3
-    exclude: xnemogcm/__init__.py
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
   rev: v0.2.0
   hooks:
     - id: ruff
+      exclude: xnemogcm/__init__.py|^.*test.*$
 
 - repo: https://github.com/codespell-project/codespell
   rev: v2.2.6
