### 2024-02-18 #155 @callumrollo
I've just about completed my review of the package. It looks to be in good shape! I particularly appreciate the comprehensive test suite. I've just raised one last issue with a missing docstring.

I don't work with nemo data, but using the example data provided I have tested that the functions in xnemogcm work as described. Hopefully @paigem can comment with some more domain specific expertise.
