### 2024-02-19 2fb2969
rm option for periodic boundaries
