### 2024-02-20 #155 @callumrollo
If this package isn't being submitted to JOSS I believe my review is complete. Thanks @rcaneill for the speedy response to Issues and PRs
