### 2024-02-29 #155 @paigem
@rcaneill I'm starting my review in the next day! Glad to see that you've been updating in real time with @callumrollo's excellent comments, since it has taken me awhile to get to my review. This is my first PyOpenSci review, so just a heads up that I may work a bit slowly (over the course of a couple days) as I learn the ropes.

@callumrollo I also have not used NEMO data before, but I have used a lot of climate model data in the past, so I'll do my best to provide domain-specific knowledge! Thanks for being prompt and thorough with your review - you're making my job slightly easier. 😉
