### 2024-03-06 rcaneill/xnemogcm#94 @paigem
Add context in documentation around how this package fits in with existing packages
**Feature suggestion:** Add more discussion about how `xnemogcm` fits into the existing suite of packages currently being used to retrieve and analyze NEMO data.

I suggest adding two sections (which could be quite brief - e.g. a sentence or two each) to the documentation home page:
1. _Put in context with existing NEMO Toolbox_
I saw buried in the description for one of the examples in the documentation that there is a NEMO Toolbox in Fortran. For context for new users, I think it would be helpful to include in the docs how `xnemogcm` compares/extends/Pythonizes/etc functionality of the existing toolbox (and of any other relevant NEMO tools).

2. _Provide some extra motivation for users to adopt this package by emphasizing easy usage with the Xarray ecosystem_
Is this the first package that allows users to open NEMO data as Xarray Datasets? If so, this should be highlighted as one of the main motivations for this package - that Xarray is widely used in ocean science, and this package will help make NEMO data more analysis-ready for those already familiar with the Xarray ecosystem. I would add this in the README or docs, as this will help new users understand why this is a useful package.
