### 2024-03-06 rcaneill/xnemogcm#96 @paigem
Add full workflow of use case example
**Feature request:** It would be helpful for users to see a full workflow of using `xnemogcm`, and seeing how it fits together with `xgcm` and how that can be used for analysis.

This is a suggested feature request, but is not one that will hold up my review. 🙂

One of the primary motivations listed for why `xnemogcm` is a useful package is that it can be used in conjunction with `xgcm`. At the moment, the only example of using `xnemogcm` with `xgcm` is in the very short "Usage" text in the README. This "Usage" text is great, but I think it would be even more helpful to have a more complete use case example that (for instance) opens up some NEMO data, uses `xnemogcm` and then uses `xgcm` to do some sort of standard ocean analysis, ideally resulting in a nice plot. Specifically, it would be nice to have a script that can be run locally, e.g. with a small subset of NEMO data (the "Usage" section as it stands cannot be run unless the user already has access to NEMO data) that demonstrates a full sample workflow.
