### 2024-03-06 #155 @paigem
@rcaneill My first reaction to this package is that it seems very useful for the scientific community, allowing those familiar with `Xarray` and `xgcm` to more easily use and analyze NEMO data. 🎉

I've been focusing on documentation today, and have a couple comments that I put into issues:
- I think it would be helpful to provide more context for how `xnemogcm` fits into the existing software ecosystem for using and analyzing NEMO data (see more detailed comments in this issue: https://github.com/rcaneill/xnemogcm/issues/94)
- The examples in the documentation could use some more descriptive language to describe what is being demonstrated. I put in a [PR](https://github.com/rcaneill/xnemogcm/pull/93) to help get started with this, but added more specific suggestions in this issue: https://github.com/rcaneill/xnemogcm/issues/95
- At the moment, the only "example" of actually using `xgcm` is in the very short "Usage" section of the README/documentation. I suggest adding a more complete sample workflow for someone wanting to analyze NEMO data. See more details in this issue: https://github.com/rcaneill/xnemogcm/issues/96
