### 2024-03-06 rcaneill/xnemogcm#94 @rcaneill
> **Feature suggestion:** Add more discussion about how `xnemogcm` fits into the existing suite of packages currently being used to retrieve and analyze NEMO data.
>
> I suggest adding two sections (which could be quite brief - e.g. a sentence or two each) to the documentation home page:
>
>   1. _Put in context with existing NEMO Toolbox_
>        I saw buried in the description for one of the examples in the documentation that there is a NEMO Toolbox in Fortran. For context for new users, I think it would be helpful to include in the docs how `xnemogcm` compares/extends/Pythonizes/etc functionality of the existing toolbox (and of any other relevant NEMO tools).

Yes, there exist tools in Fortran that are used to create domain files, input fields, etc. They are however more used to produce configurations with the necessary input files, than to analyse the outputs. So there is only one overlap, which is recombining the mesh_mask / domaincfg files, when they have been outputted by different processors.

>   2. _Provide some extra motivation for users to adopt this package by emphasizing easy usage with the Xarray ecosystem_
>        Is this the first package that allows users to open NEMO data as Xarray Datasets? If so, this should be highlighted as one of the main motivations for this package - that Xarray is widely used in ocean science, and this package will help make NEMO data more analysis-ready for those already familiar with the Xarray ecosystem. I would add this in the README or docs, as this will help new users understand why this is a useful package.

NEMO output files are outputted as netcdf so they can directly be opened by xarray. However, what is missing is all grid information in the shape needed by xgcm ([COMODO](https://web.archive.org/web/20160417032300/http://pycomodo.forge.imag.fr/norm.html) convention). To solve this problem, another python package existed when I started xnemogcm: [xorca](https://github.com/willirath/xorca). However, it is not developed any more since NEMO 3.6 version (actual NEMO version is 4.2). So xnemogcm is not the first package that allows users to open NEMO data in xarray Datasets and plug to xgcm, but it is the only one still in development and the most flexible one.

I open a PR to add some sentences in the doc.
