+++ .pre-commit-config.yaml
@@ -5,6 +5,7 @@ repos:
   rev: v4.5.0
   hooks:
     - id: trailing-whitespace
+      args: [--markdown-linebreak-ext=md]
     - id: check-ast
     - id: debug-statements
     - id: end-of-file-fixer
