### 2024-03-06 rcaneill/xnemogcm#96 @rcaneill
I wrote an example in the xgcm documentation for NEMO: https://xgcm.readthedocs.io/en/latest/xgcm-examples/04_nemo_idealized.html
Thus, I do not think that it would be necessary to redo it in the xnemogcm doc. However, I should add the link the the xgcm doc.

(I updated the xgcm example, but it is not merged yet)
