### 2024-03-06 3853566
fix trailing whitespace and update poetry
