### 2024-03-06 dd9c97f
add link to xgcm NEMO doc
