### 2024-03-06 rcaneill/xnemogcm#96 @rcaneill
(I let you reopen this issue if you think that more work would be needed)
