### 2024-03-07 5d17baf
Apply suggestions from code review
Co-authored-by: Romain Caneill <romain.caneill@ens-lyon.org>
