### 2024-03-07 rcaneill/xnemogcm#94 @paigem
Thanks for this background @rcaneill. I think it would be useful to include this background (of both the Fortran toolkit and xorca) in the README or documentation, as it is helpful for users to have this context (particularly if they have used xorca before and may not know why there is this new package).

For reference, PyOpenSci suggests including this info in the README (it's actually part of my checklist as a reviewer). 😉
