### 2024-03-07 #155 @paigem
Thanks for your quick updates @rcaneill - it all looks great!

I'm continuing to go through my review checklist and have a couple questions:
- I commented in https://github.com/rcaneill/xnemogcm/issues/94 that this background information would be helpful to include in the README.
- Do you include citation information somewhere? I don't see it anywhere at the moment, but I may have missed it. PyOpenSci suggests adding it to your README.

I am continuing through the checklist and will reach out with any further questions.
