### 2024-03-07 rcaneill/xnemogcm#94 @rcaneill
Reopening until https://github.com/rcaneill/xnemogcm/issues/94#issuecomment-1982332491 is done
