### 2024-03-07 #155 @rcaneill
I added CITATION.cff file in https://github.com/rcaneill/xnemogcm/pull/104
