+++ CITATION.cff
@@ -15,6 +15,9 @@ authors:
   - given-names: Callum
     family-names: Rollo
     orcid: 'https://orcid.org/0000-0002-5134-7886'
+  - given-names: Paige
+    family-names: Martin
+    orcid: 'https://orcid.org/0000-0003-3538-633X'
 identifiers:
   - type: doi
     value: 10.5281/zenodo.5724577
