### 2024-03-14 #155 @rcaneill
@ocefpaf I implemented / merged / commented all of their suggestions. So for me, we can move forward :)
