### 2024-04-02 #155 @ocefpaf
---

🎉 xnemogcm has been approved by pyOpenSci! Thank you Romain Caneill (@rcaneill)  for submitting xnemogcm and many thanks to Paige Martin (@paigem) and Callum Rollo (@callumrollo) for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of xnemogcm. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
