### 2024-04-12 #155 @ocefpaf
Last but not least, do you (@rcaneill, @paigem, and @callumrollo) would like to be invited to PyOS slack? It will help you share your experience, get in touch with the community, and learn from others.
