### 2024-04-15 #155 @rcaneill
* [x] Add the badge for pyOpenSci peer-review to the README.md of xnemogcm. The badge should be [![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number).
* [ ]
The new xnemogcm version is 0.4.3, the doi is: https://doi.org/10.5281/zenodo.10973846

Thanks again @ocefpaf , @callumrollo, and @paigem for the review
