### 2024-04-15 #155 @rcaneill
> Last but not least, do you (@rcaneill, @paigem, and @callumrollo) would like to be invited to PyOS slack? It will help you share your experience, get in touch with the community, and learn from others.

Thanks for the invite, but I prefer to not use Slack. I am still available by email for future exchanges.
