+++ .github/workflows/publish.yml
@@ -9,6 +9,6 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Build and publish to pypi
-        uses: JRubics/poetry-publish@v1.13
+        uses: JRubics/poetry-publish@v2.0
         with:
             pypi_token: ${{ secrets.PYPI_TOKEN }}
