### 2024-04-18 #155 @ocefpaf
> I will see if I have time for it. If so, should I open a PR on https://github.com/pyOpenSci/pyopensci.github.io?

Yes. You can see [an example here](https://github.com/pyOpenSci/pyopensci.github.io/pull/42).
